Update: able to select and create monster
