+++ client/src/components/Combat/CombatPlayerList.js
@@ -1,16 +1,18 @@
-import React, { useState} from "react";
+import React, { useState, useEffect } from "react";
 import CombatPlayer from "./CombatPlayer";
 import './CombatPlayerList.scss'
-import MonsterList from "./Monster";
+import axios from 'axios';
+import Select from 'react-select';
 
 
 
 function CombatPlayerList (props) {
 
+  const [monster, setMonster] = useState(null);
+  const [currentMonster, setCMonster] = useState(null);
   const [name, setName] = useState("");
   const [hp, setHp] = useState("");
   const [initiative, setInitiative] = useState('');
-  const [error, setError] = useState('');
   const [playerData, setPlayerData] = useState([
     {
       name: 'Lezana Carlucci',
@@ -79,29 +81,54 @@ function CombatPlayerList (props) {
       level: 7
     }
   ]);
-  
-  
 
+  //monster ======================================================
+  useEffect(() => {
+    axios.get(`https://www.dnd5eapi.co/api/monsters/`)
+    .then(res => setMonster([res.data]));
+  }, []);
+
+  const changeInput = (res) => {
+    setName(res.data.name)
+    setHp(res.data.hit_points)
+    setInitiative(res.data.dexterity)
+  }
+  const fetchingMonster = (monName) => {
+   
+    return (axios.get(`https://www.dnd5eapi.co/api/monsters/${monName}`)
+    // .then(res => setCMonster(res.data)))
+    .then (res => changeInput(res)))
+    .catch(err => console.log(err))
+ 
+  }
+  
+  function parsedMonster () {
+  
+    let options = []
+  
+    for (let i = 0; i < 332; i++) {
+      options.push ({
+        value: monster[0].results[i].index,
+        label: monster[0].results[i].name }) 
+    }
+  
+    return(<Select
+      onChange={(value) => fetchingMonster(value.value)}
+      options={options}/>)
+  }
+  
+//==========================================================================
+ 
   const reset = () => {
     setName("")
     setHp("")
-    setError('')
     setInitiative('')
   }
 
   function validate () {
-     if (name === "") {
-       setError("Name can't be empty")
-       return;
-     } 
-     if (hp === "" || hp === 0) {
-       setError("Hp can't be empty/zero")
-     }
-     if (initiative === "" || initiative === 0) {
-      setError("Initiative can't be empty/zero")
-      } else {
-      setError('');
+
     
+     
       let result = {
         name,
         stats: {
@@ -111,9 +138,10 @@ function CombatPlayerList (props) {
       };
       setPlayerData([...playerData, result]);
       reset();
-    }
+    
   }
   
+  // console.log('selected', currentMonster)
   
   const newPlayerData = playerData.sort(function(a,b) {return b.stats.initiative-a.stats.initiative});
   const parsedPlayers = newPlayerData.map(player => <CombatPlayer key={player.name} {...player}/>)
@@ -122,20 +150,24 @@ function CombatPlayerList (props) {
 return (
   <div className="CombatList">
     <section>
+      <h3>Create Your Own Character</h3>
       <form onSubmit={e => e.preventDefault()}>
         <input 
+        id="name"
         name="name"
         type='text'
         placeholder="enter your name"
         value={name}
         onChange= {e => setName(e.target.value)}/>
         <input 
+        id="hp"
         hp='hp'
         type='number'
         placeholder="enter your health"
         value={hp}
         onChange= {e => setHp(e.target.value)}/>
-        <input 
+        <input
+        id="dex"
         initiative='initiative'
         type='number'
         placeholder="enter your initiative"
@@ -143,12 +175,13 @@ return (
         onChange= {e => setInitiative(e.target.value)}/>
       </form>
       </section>
-      <section>{error}</section>
+      <h3>Pick a Monster</h3>
+      {monster !== null && parsedMonster()}
       <section>
         <button onClick={validate}>Create</button>
         <button onClick={reset}>Clear</button>
       </section>
-      <MonsterList />
+      
     {parsedPlayers}
   </div>
 )
